Renamed + final changes

+++ lab1/lab1/TVector2D.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-    public class TVector2D
-    {
+public class TVector2D
+{
     protected double x;
     protected double y;
 
@@ -19,7 +19,7 @@ using System.Threading.Tasks;
         this.y = 0;
     }
 
-    public TVector2D( TVector2D vector)
+    public TVector2D(TVector2D vector)
     {
         this.x = vector.x;
         this.y = vector.y;
@@ -28,9 +28,9 @@ using System.Threading.Tasks;
     public virtual void input()
     {
         Console.WriteLine("Please, enter x");
-        x=Int32.Parse(Console.ReadLine());
+        x = Double.Parse(Console.ReadLine());
         Console.WriteLine("Please, enter y");
-        y =Int32.Parse(Console.ReadLine());
+        y = Double.Parse(Console.ReadLine());
     }
 
     public virtual void print()
@@ -50,13 +50,13 @@ using System.Threading.Tasks;
         y = y / length;
 
     }
-    //ToDo чи треба додавати протилежність?
-    public bool equals(TVector2D vector) 
+
+    public bool equals(TVector2D vector)
     {
         return x == vector.x && y == vector.y;
     }
 
-    public static TVector2D operator + (TVector2D vector1, TVector2D vector2)
+    public static TVector2D operator +(TVector2D vector1, TVector2D vector2)
     {
         return new TVector2D(vector1.x + vector2.x, vector1.y + vector2.y);
     }
